+++ ListElab/src/app/questions/components/questions-card/questions-card.component.html
@@ -39,8 +39,8 @@
 
   <mat-menu #menu="matMenu" [matMenuTriggerFor]="menu">
     <ng-template matMenuContent>
-      <button mat-menu-item type="button" (click)="selectDetail(question)">Detalhes</button>
-      <button mat-menu-item type="button" (click)="selectDelete()">Eliminar</button>
+      <button mat-menu-item type="button" (click)="onSelect()">Detalhes</button>
+      <button mat-menu-item type="button" (click)="onDelete()">Eliminar</button>
     </ng-template>
   </mat-menu>
