+++ ListElab/src/app/questions/components/questions-card/questions-card.component.scss
