+++ ListElab/src/app/questions/questions.module.ts
@@ -7,10 +7,10 @@ import { QuestionsRoutingModule } from './questions-routing.module';
 
 import { QuestionsComponent } from './questions.component';
 
-import { QuestionsSearchComponent } from './shared/questions-search/questions-search.component';
+import { QuestionsSearchComponent } from './components/questions-search/questions-search.component';
 import { QuestionsFormComponent } from './questions-form/questions-form.component';
 import { QuestionsListComponent } from './questions-list/questions-list.component';
-import { QuestionsCardComponent } from './shared/questions-card/questions-card.component';
+import { QuestionsCardComponent } from './components/questions-card/questions-card.component';
 
 
 @NgModule({
